[Refactor] #25 - FeaturePath 분리
